tests/wpt/tests/css/css-align/parsing/row-gap-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: parsing row-gap with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-row-gap">
<meta name="assert" content="row-gap supports the full grammar 'normal | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("row-gap", "normal");

test_valid_value("row-gap", "10px");
test_valid_value("row-gap", "20%");
test_valid_value("row-gap", "calc(20% + 10px)");
test_valid_value("row-gap", "0", "0px");
</script>
</body>
</html>
